+++ src/app/[lang]/home/page.tsx
@@ -1,10 +1,11 @@
 import React from "react";
 import { getTranslation } from "../../../../locales/utils/getTranslation";
+import NavBar from "@/components/NavBar";
 
  async function Home(params: any){
     const lang = await getTranslation(params.lang);
     return(
-        <h1>{lang.hello}</h1>
+        <NavBar />
     )
 }
 export default Home;